@@ -77,53 +69,61 @@
     "jp-num": "^0.1.0",
     "jsdom": "^16.4.0",
     "jszip": "^3.5.0",
-    "kuromojin": "^2.0.0",
+    "kuromojin": "^2.1.1",
     "libmime": "^5.0.0",
     "lodash": "^4.17.20",
     "miller-rabin": "^4.0.1",
-    "moment": "^2.29.0",
+    "moment": "^2.29.1",
     "node-emoji": "^1.10.0",
     "node-persist": "^3.1.0",
     "node-schedule": "^1.3.2",
     "opentype.js": "^1.3.3",
-    "p-queue": "^6.6.1",
+    "p-queue": "^6.6.2",
     "primes-and-factors": "^1.3.3",
     "rouge": "github:kenlimmj/rouge",
     "scrape-it": "^5.2.4",
-    "sharp": "^0.26.1",
+    "sharp": "^0.27.0",
     "shogi9.js": "^1.0.2",
+    "snapsvg": "^0.5.1",
     "sql-template-strings": "^2.2.2",
-    "sqlite": "^4.0.15",
-    "sqlite3": "^5.0.0",
-    "strophe.js": "github:hakatashi/strophejs",
+    "sqlite": "^4.0.19",
+    "sqlite3": "^5.0.1",
     "suncalc": "^1.8.0",
     "trie-prefix-tree": "^1.5.1",
-    "ts-node": "^9.0.0",
+    "ts-node": "^9.1.1",
     "twitter": "^1.7.1",
-    "typescript": "^4.0.3",
-    "utf-8-validate": "^5.0.2",
+    "typescript": "^4.1.3",
+    "utf-8-validate": "^5.0.4",
     "winston": "^3.3.3",
     "winston-syslog": "^2.4.4",
-    "word2vec": "github:pizzacat83/node-word2vec",
     "xml-js": "^1.6.11",
-    "xmlhttprequest": "^1.8.0",
-    "yargs": "^16.0.3"
+    "yargs": "^16.2.0"
   },
   "devDependencies": {
-    "@hakatashi/eslint-config": "^1.13.0",
-    "@types/jest": "^26.0.14",
-    "@types/lodash": "^4.14.161",
-    "@typescript-eslint/parser": "^4.3.0",
+    "@hakatashi/eslint-config": "^1.15.0",
+    "@types/common-tags": "^1.8.0",
+    "@types/html-entities": "^1.2.16",
+    "@types/jest": "^26.0.20",
+    "@types/jsdom": "^16.2.6",
+    "@types/lodash": "^4.14.168",
+    "@types/node": "^14.14.22",
+    "@types/node-schedule": "^1.3.1",
+    "@types/opentype.js": "^1.3.1",
+    "@types/sharp": "^0.27.1",
+    "@types/twitter": "^1.7.0",
+    "@typescript-eslint/eslint-plugin": "^4.14.1",
+    "@typescript-eslint/parser": "^4.14.0",
     "byline": "^5.0.0",
-    "csv-parse": "^4.12.0",
-    "eslint": "^7.10.0",
-    "eslint-plugin-react": "^7.21.2",
-    "jest": "^26.4.2",
+    "csv-parse": "^4.15.0",
+    "eslint": "^7.18.0",
+    "eslint-formatter-rdjson": "^1.0.3",
+    "eslint-plugin-react": "^7.22.0",
+    "jest": "^26.6.3",
     "lolex": "^5.1.2",
     "mock-cloud-firestore": "^0.12.0",
-    "nodemon": "^2.0.4",
+    "nodemon": "^2.0.7",
     "prettier-eslint-cli": "^5.0.0",
-    "ts-jest": "^26.4.1"
+    "ts-jest": "^26.4.4"
   },
   "optionalDependencies": {}
 }
