@@ -3,11 +3,12 @@ module.exports = class Blocker {
 		this.blocks = new Set();
 		this.waitPromise = null;
 	}
+
 	async block(name) {
 		await this.waitPromise;
 
 		let resolve;
-		const promise = new Promise(_resolve => resolve = _resolve);
+		const promise = new Promise((_resolve) => resolve = _resolve);
 
 		const block = {name, promise, time: Date.now()};
 		this.blocks.add(block);
@@ -17,10 +18,13 @@ module.exports = class Blocker {
 			resolve();
 		};
 	}
+
 	async wait(callback, interval, intervalCallback) {
 		let intervalID;
 		if (intervalCallback) {
-			intervalID = setInterval(() => {intervalCallback(this.blocks)}, interval);
+			intervalID = setInterval(() => {
+				intervalCallback(this.blocks);
+			}, interval);
 		}
 		while (this.blocks.size > 0) {
 			await Promise.all([...this.blocks].map(({promise}) => promise));
