@@ -1,12 +1,16 @@
-// @ts-ignore
-import logger from '../lib/logger.js';
-import {get} from 'lodash';
-import {FastifyInstance} from 'fastify';
-import type {SlackInterface} from '../lib/slack';
 import {spawn} from 'child_process';
+import os from 'os';
+import {PassThrough} from 'stream';
 // @ts-ignore
 import concat from 'concat-stream';
-import {PassThrough} from 'stream';
+import {FastifyInstance} from 'fastify';
+import {get} from 'lodash';
+// @ts-ignore
+import logger from '../lib/logger.js';
+import type {SlackInterface} from '../lib/slack';
+
+// @ts-ignore
+import Blocker from './block.js';
 
 const commands = [
 	['git', 'checkout', '--', 'package.json', 'package-lock.json'],
@@ -15,23 +19,22 @@ const commands = [
 	['npm', 'install', '--production'],
 	['/home/slackbot/.cargo/bin/cargo', 'build', '--release', '--all'],
 ];
-
-// @ts-ignore
-import Blocker from './block.js';
 const deployBlocker = new Blocker();
 export const blockDeploy = (name: string) => deployBlocker.block(name);
 
+// eslint-disable-next-line require-await
 export const server = ({webClient: slack}: SlackInterface) => async (fastify: FastifyInstance) => {
 	let triggered = false;
 
 	const postMessage = (text: string) => (
 		slack.chat.postMessage({
+			username: `tsgbot-deploy [${os.hostname()}]`,
 			channel: process.env.CHANNEL_SANDBOX,
-			username: 'deploy',
 			text,
 		})
 	);
 
+	// eslint-disable-next-line require-await
 	fastify.post('/hooks/github', async (req, res) => {
 		logger.info(JSON.stringify({body: req.body, headers: req.headers}));
 
@@ -51,7 +54,9 @@ export const server = ({webClient: slack}: SlackInterface) => async (fastify: Fa
 				return 'refs not match';
 			}
 
-			if (triggered) return 'already triggered';
+			if (triggered) {
+				return 'already triggered';
+			}
 			triggered = true;
 
 			deployBlocker.wait(
@@ -66,8 +71,8 @@ export const server = ({webClient: slack}: SlackInterface) => async (fastify: Fa
 						proc.stderr.on('data', (chunk) => muxed.write(chunk));
 
 						Promise.all([
-							new Promise((resolve) => proc.stdout.on('end', () => resolve())),
-							new Promise((resolve) => proc.stderr.on('end', () => resolve())),
+							new Promise<void>((resolve) => proc.stdout.on('end', () => resolve())),
+							new Promise<void>((resolve) => proc.stderr.on('end', () => resolve())),
 						]).then(() => {
 							muxed.end();
 						});
@@ -84,18 +89,18 @@ export const server = ({webClient: slack}: SlackInterface) => async (fastify: Fa
 
 					await postMessage('死にます:wave:');
 
-					await new Promise(resolve =>
-						setTimeout(() => {
-							process.exit(0);
-							resolve();
-						}, 2000)
-					);
+					await new Promise<void>((resolve) => setTimeout(() => {
+						// eslint-disable-next-line no-process-exit, node/no-process-exit
+						process.exit(0);
+						// eslint-disable-next-line no-unreachable
+						resolve();
+					}, 2000));
 				},
 				30 * 60 * 1000, // 30min
 				(blocks: any) => {
 					logger.info(blocks);
 					postMessage('デプロイがブロック中だよ:confounded:');
-				}
+				},
 			);
 
 			return 'ok';
