@@ -6,14 +6,14 @@ import axios from 'axios';
 import cheerio from 'cheerio';
 // @ts-ignore
 import levenshtein from 'fast-levenshtein';
-import {AllHtmlEntities} from 'html-entities';
+import {google} from 'googleapis';
+import {decode as decodeHtmlEntities} from 'html-entities';
 import iconv from 'iconv-lite';
 // @ts-ignore
 import {hiraganize} from 'japanese';
 import {random, sample, shuffle, flatten, times, constant, range} from 'lodash';
 import scrapeIt from 'scrape-it';
 import type {SlackInterface} from '../lib/slack';
-import {google} from 'googleapis';
 
 const mutex = new Mutex();
 
@@ -21,6 +21,7 @@ export interface Quiz {
 	id: number,
 	question: string,
 	answer: string,
+	author?: string,
 }
 
 interface Data {
@@ -29,12 +30,13 @@ interface Data {
 
 const statePath = path.resolve(__dirname, 'candidates.json');
 const itStatePath = path.resolve(__dirname, 'candidates-it.json');
+const hakatashiItStatePath = path.resolve(__dirname, 'candidates-hakatashi-it.json');
 
 const fullwidth2halfwidth = (string: string) => (
 	string.replace(/[\uFF01-\uFF5E]/gu, (char) => String.fromCodePoint(char.codePointAt(0) - 0xFF00 + 0x20))
 );
 
-export const normalize = (string: string) => { 
+export const normalize = (string: string) => {
 	let newString = string;
 	newString = newString.replace(/\(.+?\)/g, '');
 	newString = newString.replace(/\[.+?\]/g, '');
@@ -42,7 +44,7 @@ export const normalize = (string: string) => {
 	newString = newString.replace(/【.+?】/g, '');
 	newString = newString.replace(/[^\p{Letter}\p{Number}]/gu, '');
 	newString = newString.toLowerCase();
-	return hiraganize(fullwidth2halfwidth(newString))
+	return hiraganize(fullwidth2halfwidth(newString));
 };
 
 const getQuiz = async () => {
@@ -93,13 +95,12 @@ export const getHardQuiz = async () => {
 		formname: 'lite_search',
 	})}`;
 
-	const entities = new AllHtmlEntities();
 	const {data: quiz} = await scrapeIt<Quiz>(url, {
 		id: 'tbody td:nth-child(1)',
 		question: {
 			selector: 'tbody td:nth-child(3) > a',
 			how: 'html',
-			convert: (x) => entities.decode(x),
+			convert: (x) => decodeHtmlEntities(x),
 		},
 		answer: 'tbody td:nth-child(4)',
 	});
@@ -153,6 +154,52 @@ export const getItQuiz = async () => {
 	const quiz = quizes.find((q) => q.id === id);
 
 	await fs.writeFile(itStatePath, JSON.stringify(candidates.filter((candidate) => candidate !== id)));
+
+	return quiz;
+};
+
+export const getHakatashiItQuiz = async () => {
+	const stateExists = await fs.access(hakatashiItStatePath, constants.F_OK).then(() => true).catch(() => false);
+	const candidates = [];
+	if (stateExists) {
+		const stateData = await fs.readFile(hakatashiItStatePath);
+		candidates.push(...JSON.parse(stateData.toString()));
+	}
+
+	if (candidates.length === 0) {
+		candidates.push(...range(1, 450));
+	}
+
+	const id = sample(candidates);
+
+	const auth = await new google.auth.GoogleAuth({
+		scopes: ['https://www.googleapis.com/auth/spreadsheets.readonly'],
+	}).getClient();
+
+	const sheets = google.sheets({version: 'v4', auth});
+
+	const values: [string, string, string][] = await new Promise((resolve, reject) => {
+		sheets.spreadsheets.values.get({
+			spreadsheetId: '1357WnNdRvBlDnh3oDtIde7ptDjm2pFFFb-hbytFX4lk',
+			range: 'hakatashi!A:C',
+		}, (error, response) => {
+			if (error) {
+				reject(error);
+			} else if (response.data.values) {
+				resolve(response.data.values as [string, string, string][]);
+			} else {
+				reject(new Error('values not found'));
+			}
+		});
+	});
+
+	const quizes: Quiz[] = values.map(([id, question, answer]) => ({
+		id: parseInt(id), question, answer, author: '320061621395259392',
+	}));
+
+	const quiz = quizes.find((q) => q.id === id);
+
+	await fs.writeFile(hakatashiItStatePath, JSON.stringify(candidates.filter((candidate) => candidate !== id)));
 
 	return quiz;
 };
