@@ -132,14 +132,19 @@ async function getDictionary() {
 module.exports = (clients) => {
 	const { rtmClient: rtm, webClient: slack } = clients;
 
-	function postMessage(message, channel, broadcast = false) {
+	function postMessage(message, channel, postThreadOptions = {}) {
+		const {broadcast, threadPosted} = {
+			broadcast: false,
+			threadPosted: null,
+			...postThreadOptions,
+		}
 		return slack.chat.postMessage({
 			channel,
 			text: message,
 			as_user: false,
 			username: "pocky",
 			icon_emoji: ":google:",
-			thread_ts: thread,
+			thread_ts: threadPosted ? threadPosted : null,
 			reply_broadcast: broadcast,
 		});
 	}
@@ -179,21 +184,21 @@ module.exports = (clients) => {
 			スレッドで回答してね!
 
 			${hints.map((hint) => hint.replace(theme.word, '• 〇〇')).join('\n')}
-		`, process.env.CHANNEL_SANDBOX);
+		`, process.env.CHANNEL_SANDBOX, {broadcast: false});
 
 		thread = message.ts;
 
-		await postMessage('3分経過で答えを発表するよ～', process.env.CHANNEL_SANDBOX);
+		await postMessage('3分経過で答えを発表するよ～', process.env.CHANNEL_SANDBOX, {broadcast: false, threadPosted: thread});
 		const currentTheme = theme;
 		setTimeout(async () => {
 			if (theme === currentTheme) {
 				await postMessage(stripIndents`
 					なんでわからないの？
 					答えは＊${theme.word}＊ (${theme.ruby}) だよ:anger:
-				`, process.env.CHANNEL_SANDBOX, true);
+				`, process.env.CHANNEL_SANDBOX, {broadcast: true, threadPosted: thread});
 				await postMessage(stripIndents`
 					${hints.map((hint) => hint.replace(theme.word, `• ＊${theme.word}＊`)).join('\n')}
-				`, process.env.CHANNEL_SANDBOX);
+				`, process.env.CHANNEL_SANDBOX, {broadcast: false, threadPosted: thread});
 				theme = null;
 				thread = null;
 			}
@@ -213,10 +218,10 @@ module.exports = (clients) => {
 				await postMessage(stripIndents`
 					<@${message.user}> 正解:tada:
 					答えは＊${word}＊ (${ruby}) だよ:tada:
-				`, channel, true);
+				`, channel, {broadcast: true, threadPosted: thread});
 				await postMessage(stripIndents`
 					${hints.map((hint) => hint.replace(word, `• ＊${word}＊`)).join('\n')}
-				`, channel);
+				`, channel, {broadcast: false, threadPosted: thread});
 				increment(message.user, "pockygame-win");
 				const date = new Date().toLocaleString('en-US', {
 					timeZone: 'Asia/Tokyo',
@@ -251,7 +256,7 @@ module.exports = (clients) => {
 		}
 		const result = await reply(match[1], match[2].length - 1);
 		if (result !== null) {
-			postMessage(htmlEscape(result), channel);
+			postMessage(htmlEscape(result), channel, {broadcast: false, threadPosted: thread_ts});
 			unlock(message.user, "pocky");
 			if (Array.from(result).length >= 20) {
 				unlock(message.user, "long-pocky");
