@@ -59,7 +59,7 @@ describe('slacklog', () => {
 	});
 
 	it('respond to slack hook of slacklog unfurling', async () => {
-		const done = new Promise((resolve) => {
+		const done = new Promise<void>((resolve) => {
 			// @ts-ignore
 			axios.mockImplementation(({url, data}: {url: string, data: any}) => {
 				if (url === 'https://slack.com/api/chat.unfurl') {
